Create plus one
This is a basic thought which will come in mind that we can have only two cases:

Our last digit is not 9
Our last digit is 9
In case 1, it is very easy to solve as we just have to increment the number by one and give the resulting vector as the output.
In case 2, we will have to do some task as when we increment 9 it will give carry one to it's left number. So, to take care of this situation we used a while loop. This loop will run till either we have covered the whole vector in which case i becomes -1 or our last number is no longer 9.
If we have run through the whole vector then we first need to push 1 to our result vector and then we directly push every element of our modified vector.
If our last number is not 9 then we simply increment that number and then push this modified vector to our result vector